Fixed Bug Causing SendTransport to Send Messages Serially
Sending a request with SendTransport required a `&mut` reference to
`SendTransport`. This meant only one thing could hold a reference and
make requests at a time.

A mutual exclusion thing is needed to ensure that only one send is
started at a time (UdpSocketSendHalf), but many requests can be
in-flight awaiting a response at a time.

Moved the mutex inside SendTransport and made methods to send requests
use non-mutable references. This one send to happen at a time while still
allowing multiple requests to be in flight.

Now 100% CPU utilization can easily be hit for this I/O bound workload.